Add all components for building interceptor image
Added dockerfile and cmakelists for building interceptor image for car's collision control. Also put all components in dedicated folder
