+++ src/Interceptor/README.md
@@ -1,0 +1,3 @@
+# Interceptor
+
+This folder contains components needed for building Interceptor docker image for car's collision control
